refactor: reorganize selection context and related components for improved group row selection handling
