+++ React/src/GroupRowSelection/selection-context/helpers.ts
