+++ React/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
-import { GroupRowSelectionProvider } from "./GroupRowSelection/context/GroupRowSelectionContext.tsx";
+import { GroupRowSelectionProvider } from "./GroupRowSelection/selection-context/row-selection-context.tsx";
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
